requirements.txt
django
black
flake8
python-dotenv
pytest-django
celery
redis